Work on Rest handlers, some repo work, HttpClient and ApiClient for Untappd work. Adapter between Untappd and application object work. Investigation to Authenticating against Untappd Api (delaying/putting off for now). Diving into dynamic objects for handling content from Untappd's Api. Some DI work/architecting/investigation for the Untappd Authentication.

+++ CMSC495Team3ServerApp/Models/Untappd/Beer.cs
@@ -5,7 +5,7 @@
         public int bid { get; set; }
         public string beer_name { get; set; }
         public string beer_label { get; set; }
-        public double beer_abv { get; set; }
+        public float beer_abv { get; set; }
         public int beer_ibu { get; set; }
         public string beer_description { get; set; }
         public string beer_style { get; set; }
